# backend/docs/swagger.yaml
basePath: /api
definitions:
  handlers.avatarResponse:
    properties:
      age:
        type: integer
      baseVideoS3Key:
        type: string
      baseVideoS3Url:
        type: string
      category:
        type: string
      createdAt:
        type: string
      ethnicity:
        type: string
      heightCm:
        type: integer
      id:
        type: integer
      imageS3Key:
        type: string
      imageS3Url:
        type: string
      initQueuePos:
        type: integer
      liveSettings:
        $ref: '#/definitions/models.LiveSettings'
      name:
        type: string
      personality:
        type: string
      relationshipStatus:
        type: string
      status:
        type: string
      voiceId:
        type: string
      weightKg:
        type: integer
    type: object
...
  models.LiveSettings:
    properties:
      subtitleBorder:
        description: stroke width in px (0 = none)
        type: integer
      subtitleEnabled:
        type: boolean
      subtitleFont:
        description: font filename in worker/fonts/, "" = system default
        type: string
      subtitlePosition:
        description: '"bottom" | "top"'
        type: string
      subtitleSize:
        description: px
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: LingCast 数字人直播平台 API（api-admin / api-user / api-scheduler 共用一份 OpenAPI
    文档；管理端写操作与 /users、/admin/* 需登录会话，观众端与 Worker Webhook 公开）。
  title: LingCast API
  version: 1.0.0
paths:
...
  /avatars:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List all avatars
      tags:
      - avatars
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar name
        in: formData
        name: name
        required: true
        type: string
      - description: Appearance image
        in: formData
        name: image
        required: true
        type: file
      - description: Edge-TTS voice id
        in: formData
        name: voice_id
        type: string
      - description: Category (闲聊/知识/娱乐/游戏/带货/其他)
        in: formData
        name: category
        type: string
      - description: Age
        in: formData
        name: age
        type: integer
      - description: Height in cm
        in: formData
        name: height_cm
        type: integer
      - description: Weight in kg
        in: formData
        name: weight_kg
        type: integer
      - description: Ethnicity
        in: formData
        name: ethnicity
        type: string
      - description: Relationship status
        in: formData
        name: relationship_status
        type: string
      - description: Personality
        in: formData
        name: personality
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.avatarResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 'Create an avatar (multipart: name + image + optional persona fields)'
      tags:
      - avatars
  /avatars/{id}:
    delete:
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete an avatar (cascades tasks/sessions/files)
      tags:
      - avatars
    get:
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.avatarResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a single avatar (includes liveSettings)
      tags:
      - avatars
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.avatarResponse'
      summary: Edit an avatar (name/category/voice/persona profile)
      tags:
      - avatars
  /avatars/{id}/base-video:
    post:
      consumes:
      - application/json
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      - description: baseVideoS3Key + status (+ coverS3Key)
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.avatarResponse'
      summary: 'Worker webhook: persist base video + status'
      tags:
      - worker
...
  /avatars/{id}/retry:
    post:
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Regenerate the base driving video
      tags:
      - avatars
  /avatars/{id}/skip:
    post:
      parameters:
      - description: Avatar ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Skip base-video generation for an avatar
      tags:
      - avatars
...
swagger: "2.0"